scripts/scraper/requirements.txt
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncpg>=0.29.0
certifi>=2024.0.0
